refactors code to manage Bookmark collection state in AppContext
