+++ src/components/Bookmark.jsx
@@ -15,95 +15,102 @@ import {
 import bookmarkIcon from "../assets/icons/bookmark.svg";
 import { useAppContext } from "../AppContext";
 
-const HARD_CODED_CATEGORIES = [
-    {
-        id: "ro7Sz05bCKdfzFaYUOx7",
-        name: "network",
-        userID: "dNC63cyuDbEoEntxBpe9",
-        contentID: [
-            "HIM6R8AbiEKBZWhkIy8Y",
-            "j9Tq3xCdLp7mRv1sFg0H",
-            "0c4a01729ebdc11d608865176acd925ce0625353fa6c60982c284e16bd4eefb9",
-        ],
-        createdAt: "July 3, 2024 at 1:55:50 AM UTC-4",
-    },
-    {
-        id: "lRqX0IFdr6u1gQXRBGa1",
-        name: "drive",
-        userID: "dNC63cyuDbEoEntxBpe9",
-        contentID: [
-            "afYzXislW1iopWhNyQF3",
-            "07cfdd3433077bf9e3b11b15a41e1535c0609342b731a59f573044905b2997d0",
-        ],
-        createdAt: "March 6, 2024 at 12:32:25 AM UTC-5",
-    },
-];
+// const HARD_CODED_CATEGORIES = [
+//     {
+//         id: "ro7Sz05bCKdfzFaYUOx7",
+//         name: "network",
+//         userID: "dNC63cyuDbEoEntxBpe9",
+//         contentID: [
+//             "HIM6R8AbiEKBZWhkIy8Y",
+//             "j9Tq3xCdLp7mRv1sFg0H",
+//             "0c4a01729ebdc11d608865176acd925ce0625353fa6c60982c284e16bd4eefb9",
+//         ],
+//         createdAt: "July 3, 2024 at 1:55:50 AM UTC-4",
+//     },
+//     {
+//         id: "lRqX0IFdr6u1gQXRBGa1",
+//         name: "drive",
+//         userID: "dNC63cyuDbEoEntxBpe9",
+//         contentID: [
+//             "afYzXislW1iopWhNyQF3",
+//             "07cfdd3433077bf9e3b11b15a41e1535c0609342b731a59f573044905b2997d0",
+//         ],
+//         createdAt: "March 6, 2024 at 12:32:25 AM UTC-5",
+//     },
+// ];
 
-export default function Bookmark({ contentId }) {
-    console.log(`content Id being looked at: ${contentId}`);
+export default function Bookmark({ contentInfo }) {
     /*
     Bookmark should have a prop contentInfo:
         {contentId, userId?} = contentInfo
     */
     const [isDialogVisible, setIsDialogVisible] = useState(false);
     const [isDialogOpen, setIsDialogOpen] = useState(false);
-    const [categories, setCategories] = useState([]);
+    const {
+        content,
+        currentIndex,
+        categories,
+        handleCategoryToggle,
+        handleCreateNewCollection,
+    } = useAppContext();
+    // const [categories, setCategories] = useState([]);
+    console.log(`content Id being looked at: ${contentInfo.id}`);
 
-    function handleCategoryToggle(categoryIndex) {
-        /*
-        if selectedCategories.includes(categoryId)
-            then it was already checked so uncheck it now...
-            -remove contentId from category.contentID...use array filter 
-            -UPDATE category collection in DB using categoryId
-            -if DB update successful, display confimation message
-                -else display error message
-        else
-            it was unchecked so check it
-            -add contentId to category.contentID array
-            -update category collection in db 
-            -if DB update successful, display confimation method
-                -else display error message
+    // function handleCategoryToggle(categoryIndex) {
+    //     /*
+    //     if selectedCategories.includes(categoryId)
+    //         then it was already checked so uncheck it now...
+    //         -remove contentId from category.contentID...use array filter
+    //         -UPDATE category collection in DB using categoryId
+    //         -if DB update successful, display confimation message
+    //             -else display error message
+    //     else
+    //         it was unchecked so check it
+    //         -add contentId to category.contentID array
+    //         -update category collection in db
+    //         -if DB update successful, display confimation method
+    //             -else display error message
 
-         */
-        const newCategories = [...categories];
-        const indexOfID = newCategories[categoryIndex].contentID.findIndex(
-            (id) => id === contentId
-        );
-        indexOfID !== -1
-            ? newCategories[categoryIndex].contentID.splice(indexOfID, 1)
-            : newCategories[categoryIndex].contentID.push(contentId);
+    //      */
+    //     const newCategories = [...categories];
+    //     const indexOfID = newCategories[categoryIndex].contentID.findIndex(
+    //         (id) => id === contentId
+    //     );
+    //     indexOfID !== -1
+    //         ? newCategories[categoryIndex].contentID.splice(indexOfID, 1)
+    //         : newCategories[categoryIndex].contentID.push(contentId);
 
-        setCategories(newCategories);
-    }
+    //     setCategories(newCategories);
+    // }
 
-    for (let cat of categories) {
-        console.log(`${cat.name} --content ids: ${cat.contentID}`);
-    }
+    // for (let cat of categories) {
+    //     console.log(`${cat.name} --content ids: ${cat.contentID}`);
+    // }
 
-    function handleCreateNewCollection(newCategoryInfo) {
-        /*
-        -use contentId prop to create newCategory object
-            -const {name, createdAt} = newCategoryInfo
-            -create a random category Id that look like ids in DB
-            -format createdAt date to look like data in DB**
-            -where to get userID????
+    // function handleCreateNewCollection(newCategoryInfo) {
+    //     /*
+    //     -use contentId prop to create newCategory object
+    //         -const {name, createdAt} = newCategoryInfo
+    //         -create a random category Id that look like ids in DB
+    //         -format createdAt date to look like data in DB**
+    //         -where to get userID????
 
-        -what about the icon picture? logic for how to get it?
-        
-        -setCategories
-        -display confirmation message (inside NewCollectionPopup dialog, before removing visiability)
-        -POST To firestore DB
-        
-         */
-        let newCategory = {
-            id: Date.now(),
-            userID: "dNC63cyuDbEoEntxBpe9",
-            contentID: [],
-            ...newCategoryInfo,
-        };
-        console.log(newCategory);
-        setCategories([...categories, newCategory]);
-    }
+    //     -what about the icon picture? logic for how to get it?
+
+    //     -setCategories
+    //     -display confirmation message (inside NewCollectionPopup dialog, before removing visiability)
+    //     -POST To firestore DB
+
+    //      */
+    //     let newCategory = {
+    //         id: Date.now(),
+    //         userID: "dNC63cyuDbEoEntxBpe9",
+    //         contentID: [],
+    //         ...newCategoryInfo,
+    //     };
+    //     console.log(newCategory);
+    //     setCategories([...categories, newCategory]);
+    // }
 
     function handleCloseDialog() {
         setIsDialogVisible(false);
@@ -121,15 +128,14 @@ export default function Bookmark({ contentId }) {
         }
     }, [isDialogOpen]);
 
-    useEffect(() => {
-        console.log("inside useEffect");
-        /**Add code here to fetch Categories from DB by doing:
-         * import {fetchUsersCategories, fetchCategoriesByUserId} from ../functions
-         *
-         * result from the function should give back an array
-         */
-        setCategories(HARD_CODED_CATEGORIES);
-    }, []);
+    // useEffect(() => {
+    //     console.log("inside useEffect");
+    //     /**Add code here to fetch Categories from DB by doing:
+    //      * import {fetchUsersCategories, fetchCategoriesByUserId} from ../functions
+    //      * result from the function should give back an array
+    //      */
+    //     setCategories(HARD_CODED_CATEGORIES);
+    // }, []);
 
     return (
         <>
@@ -177,10 +183,13 @@ export default function Bookmark({ contentId }) {
                                                 //     category.id
                                                 // )}
                                                 checked={category.contentID.includes(
-                                                    contentId
+                                                    contentInfo.id
                                                 )}
                                                 onCheckedChange={() =>
-                                                    handleCategoryToggle(index)
+                                                    handleCategoryToggle(
+                                                        index,
+                                                        contentInfo.id
+                                                    )
                                                 }
                                                 className="border-[#546672] peer data-[state=checked]:bg-[#0264D4]"
                                             />
